Add files via upload
Program made to calculate the costs of a paint job for a fictional company. 

A painting company has determined that for every 350 square feet of wall space, one gallon of paint and six hours of labor are required.  The company charges $62.25 per hour for labor. The program should ask the user to enter the square feet of wall space to be painted and the price of the paint per gallon. The program is to display the following information.

-The number of gallons of paint required rounded up to whole gallons.
-The hours of labor required.
-The cost of the paint based on the rounded up whole gallons.
-The labor charges.
-The total cost of the paint job.